Move SubscriptionUuid to SeederFolder
