+++ database/seeders/SubscriptionUuid.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Domain\Subscriptions;
+namespace Database\Seeders;
 
 enum SubscriptionUuid: string
 {
